+++ src/Constants.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace LaLit;
+
+class Constants
+{
+    const DEFAULT_DOM_VERSION = '1.0';
+    const DEFAULT_ENCODING = 'utf-8';
+    const DEFAULT_STANDALONE = false;
+    const DEFAULT_FORMAT_OUTPUT = true;
+
+    const LABEL_ATTRIBUTES = '@attributes';
+    const LABEL_CDATA = '@cdata';
+    const LABEL_DOCTYPE = '@docType';
+    const LABEL_VALUE = '@value';
+}